+++ stream-compositions/services/transaction-composition-service/src/main/java/com/backbase/stream/compositions/transaction/core/service/impl/TransactionIngestionServiceImpl.java
@@ -1,7 +1,7 @@
 package com.backbase.stream.compositions.transaction.core.service.impl;
 
-import com.backbase.dbs.transaction.api.service.v2.model.TransactionsPostRequestBody;
-import com.backbase.dbs.transaction.api.service.v2.model.TransactionsPostResponseBody;
+import com.backbase.dbs.transaction.api.service.v3.model.TransactionsPostRequestBody;
+import com.backbase.dbs.transaction.api.service.v3.model.TransactionsPostResponseBody;
 import com.backbase.stream.TransactionService;
 import com.backbase.stream.compositions.transaction.core.config.TransactionConfigurationProperties;
 import com.backbase.stream.compositions.transaction.core.mapper.TransactionMapper;
@@ -197,8 +197,7 @@ public class TransactionIngestionServiceImpl implements TransactionIngestionServ
      * @return Ingested transactions
      */
     private Mono<List<TransactionsPostResponseBody>> sendToDbs(Flux<TransactionsPostRequestBody> transactions) {
-        return transactions
-                .publish(transactionService::processTransactions)
+        return transactionService.processTransactions(transactions)
                 .flatMapIterable(UnitOfWork::getStreamTasks)
                 .flatMapIterable(TransactionTask::getResponse)
                 .collectList();